Move recordmethod.h to be an "internal" header
