+++ ssl/record/record.h
@@ -10,7 +10,7 @@
 typedef struct ssl_connection_st SSL_CONNECTION;
 
 #include <openssl/core_dispatch.h>
-#include "recordmethod.h"
+#include "internal/recordmethod.h"
 
 /*****************************************************************************
  *                                                                           *
