+++ test/sslbuffertest.c
@@ -16,7 +16,7 @@
 /* We include internal headers so we can check if the buffers are allocated */
 #include "../ssl/ssl_local.h"
 #include "../ssl/record/record_local.h"
-#include "../ssl/record/recordmethod.h"
+#include "internal/recordmethod.h"
 #include "../ssl/record/methods/recmethod_local.h"
 
 #include "internal/packet.h"
