+++ test/tls13encryptiontest.c
@@ -11,7 +11,7 @@
 #include <openssl/evp.h>
 #include "../ssl/ssl_local.h"
 #include "../ssl/record/record_local.h"
-#include "../ssl/record/recordmethod.h"
+#include "internal/recordmethod.h"
 #include "../ssl/record/methods/recmethod_local.h"
 #include "internal/nelem.h"
 #include "testutil.h"
